requirementss.txt

python-dotenv
numpy

faiss-cpu==1.7.4
google-genai
ollama==0.3.1

# LangChain
langchain==0.1.20
langchain-community==0.0.38

# Web frameworks
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0

# File processing
PyPDF2==3.0.1
python-multipart==0.0.6

# Data processing
pandas==2.0.3

# Development
pytest==7.4.3